新增功能(app): 数据库资金类型由 `Float` 修改为 `Double`；
重大改变: 无

本次解决的问题: 无

+++ app/src/main/kotlin/cn/wj/android/cashbook/data/database/table/AssetTable.kt
@@ -34,7 +34,7 @@ data class AssetTable(
     @PrimaryKey(autoGenerate = true) val id: Long?,
     @ColumnInfo(name = "books_id") val booksId: Long,
     val name: String,
-    @ColumnInfo(name = "total_amount") val totalAmount: Float,
+    @ColumnInfo(name = "total_amount") val totalAmount: Double,
     @ColumnInfo(name = "billing_date") val billingDate: String,
     @ColumnInfo(name = "repayment_date") val repaymentDate: String,
     val type: String,
